fixed network checks once and for all
git-svn-id: http://aptoide.org/repo/aptoide-client/unstable/Aptoide@276 8d55eb69-6847-4b75-98a7-935b3913b0ec

+++ src/cm/aptoide/pt/DbHandler.java
@@ -19,9 +19,16 @@
 
 package cm.aptoide.pt;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Vector;
+
+import cm.aptoide.pt.multiversion.VersionApk;
+
+
+
+
 
 import android.content.ContentValues;
 import android.content.Context;
@@ -45,8 +52,11 @@ public class DbHandler {
 	private static final String TABLE_NAME_LOCAL = "local";
 	private static final String TABLE_NAME = "aptoide";
 	private static final String TABLE_NAME_URI = "servers";
-	
 	private static final String TABLE_NAME_EXTRA = "extra";
+	private static final String TABLE_NAME_OLD_VERSIONS = "old_versions";
+	
+	
+	
 	
 	private static SQLiteDatabase db = null;
 	
@@ -54,8 +64,8 @@ public class DbHandler {
 	            + "name text not null, path text not null, lastver text not null, lastvercode number not null, "
 	            + "server text, md5hash text, size number default 0 not null, primary key(apkid, server));";
 	
-	private static final String CREATE_TABLE_LOCAL = "create table if not exists " + TABLE_NAME_LOCAL + " (apkid text primary key, "
-				+ "instver text not null, instvercode number not null);";
+	private static final String CREATE_TABLE_LOCAL = "create table if not exists " + TABLE_NAME_LOCAL + " (apkid text, "
+				+ "instver text not null, instvercode number not null, primary key(apkid));";
 	
 	private static final String CREATE_TABLE_URI = "create table if not exists " + TABLE_NAME_URI 
 				+ " (uri text primary key, inuse integer not null, napk integer default -1 not null, user text, psd text,"
@@ -65,15 +75,19 @@ public class DbHandler {
 				+ " (apkid text, rat number, dt date, desc text, dwn number, catg text default 'Other' not null,"
 				+ " catg_ord integer default 2 not null, primary key(apkid));";
 	
+	private static final String CREATE_TABLE_OLD_VERSIONS = "create table if not exists " + TABLE_NAME_OLD_VERSIONS
+				+ " (apkid text, name text not null, path text not null, ver text not null, vercode number not null, " 
+				+ " md5hash text, size number default 0 not null, server text, primary key(apkid,ver,server));";
 	
 	
 	Map<String, Object> getCountSecCatg(int ord){
 		final String basic_query = "select a.catg, count(a.apkid) from " + TABLE_NAME_EXTRA + " as a where a.catg_ord = " + ord + " and not exists" +
 								   " (select * from " + TABLE_NAME_LOCAL + " as b where b.apkid = a.apkid) group by catg;"; 
+		
 		//final String basic_query2 = "select catg, count(*) from " + TABLE_NAME_EXTRA + " where catg_ord = " + ord + " group by catg;";
 		Map<String, Object> count_lst = new HashMap<String, Object>();
 		Cursor q = null;
-
+		
 		q = db.rawQuery(basic_query, null);
 		if(q.moveToFirst()){
 			count_lst.put(q.getString(0), q.getInt(1));
@@ -109,9 +123,6 @@ public class DbHandler {
 		}
 	}
 	
-	
-	
-	
 	/*
 	 * catg_ord: game (0) / application (1) / others(2) 
 	 * 
@@ -130,6 +141,7 @@ public class DbHandler {
 			db.execSQL(CREATE_TABLE_EXTRA);
 			db.execSQL(CREATE_TABLE_APTOIDE);
 			db.execSQL(CREATE_TABLE_LOCAL);
+			db.execSQL(CREATE_TABLE_OLD_VERSIONS);
 		}else if(!db.isOpen()){
 			db = ctx.openOrCreateDatabase(DATABASE_NAME, 0, null);
 		}
@@ -202,11 +214,12 @@ public class DbHandler {
 			db.execSQL("drop table if exists " + TABLE_NAME_EXTRA);
 			db.execSQL("drop table if exists " + TABLE_NAME_LOCAL);
 			db.execSQL("drop table if exists " + TABLE_NAME_URI);
+			db.execSQL("drop table if exists " + TABLE_NAME_OLD_VERSIONS);
 			db.execSQL(CREATE_TABLE_URI);
 			db.execSQL(CREATE_TABLE_EXTRA);
 			db.execSQL(CREATE_TABLE_APTOIDE);
 			db.execSQL(CREATE_TABLE_LOCAL);
-			
+			db.execSQL(CREATE_TABLE_OLD_VERSIONS);
 			/*for(String uri: repos){
 				ContentValues tmp = new ContentValues();
 				tmp.put("uri", uri);
@@ -230,11 +243,12 @@ public class DbHandler {
 		}catch(Exception e){ }
 	}
 	
-	public void delApk(String apkid){
-		db.delete(TABLE_NAME, "apkid='"+apkid+"'", null);
-		db.delete(TABLE_NAME_EXTRA, "apkid='"+apkid+"'", null);
+	public void delApk(String apkid, String ver){
+		if(db.delete(TABLE_NAME, "apkid='"+apkid+"' and lastver='"+ver+"'", null)==1)
+			db.delete(TABLE_NAME_EXTRA, "apkid='"+apkid+"'", null);
+		db.delete(TABLE_NAME_OLD_VERSIONS, "apkid='"+apkid+"' and ver='"+ver+"'", null);
 	}
-	
+	 
 	public void insertApk(boolean delfirst, String name, String path, String ver, int vercode ,String apkid, String date, Float rat, String serv, String md5hash, int down, String catg, int catg_type, int size){
 
 		if(delfirst){
@@ -271,6 +285,43 @@ public class DbHandler {
 			PackageInfo pkginfo = mPm.getPackageInfo(apkid, 0);
 			String vers = pkginfo.versionName;
 		    int verscode = pkginfo.versionCode;
+		    insertInstalled(apkid, vers, verscode);
+		} catch (NameNotFoundException e) {
+			//Not installed... do nothing
+		}
+		
+	}
+	
+	/**
+	 * @author rafael
+	 * 
+	 * @param name
+	 * @param path
+	 * @param ver
+	 * @param vercode
+	 * @param apkid
+	 * @param serv
+	 * @param md5hash
+	 * @param size
+	 */
+	public void insertOldApk(String name, String path, String ver, int vercode ,String apkid, String serv, String md5hash, int size){
+		
+		ContentValues tmp = new ContentValues();
+		tmp.put("apkid", apkid);
+		tmp.put("name", name);
+		tmp.put("path", path);
+		tmp.put("ver", ver);
+		tmp.put("vercode", vercode);
+		tmp.put("md5hash", md5hash);
+		tmp.put("size", size);
+		tmp.put("server", serv);
+		db.insert(TABLE_NAME_OLD_VERSIONS, null, tmp);
+		
+   		PackageManager mPm = mctx.getPackageManager();
+		try {
+			PackageInfo pkginfo = mPm.getPackageInfo(apkid, 0);
+			String vers = pkginfo.versionName;
+		    int verscode = pkginfo.versionCode;
 			insertInstalled(apkid, vers, verscode);
 		} catch (NameNotFoundException e) {
 			//Not installed... do nothing
@@ -278,25 +329,53 @@ public class DbHandler {
 		
 	}
 	
-	public boolean insertInstalled(String apkid){
+	/**
+	 * @author rafael
+	 * 
+	 * @param tmp
+	 * @param mserver
+	 */
+	public void insertOldApk(ApkNodeFull tmp, String mserver){
+		insertOldApk(tmp.name, tmp.path, tmp.ver, tmp.vercode,tmp.apkid, mserver, tmp.md5hash, tmp.size);
+	}
+	
+	/**
+	 * @author rafael
+	 * 
+	 * 
+	 * @param apkid
+	 * @param ver
+	 * @return
+	 */
+	public boolean insertInstalled(String apkid, String ver){
 		ContentValues tmp = new ContentValues();
 		tmp.put("apkid", apkid);
-		Cursor c = db.query(TABLE_NAME, new String[] {"lastver", "lastvercode"}, "apkid=\""+apkid+"\"", null, null, null, null);
-		c.moveToFirst();
-		String ver = c.getString(0);
-		int vercode = c.getInt(1);
+		Cursor c = db.query(TABLE_NAME, new String[] {"lastvercode"}, " apkid=\""+apkid+"\" and lastver=\""+ver+"\" ", null, null, null, null);
+		
+		if(!c.moveToFirst()){
+			c = db.query(TABLE_NAME_OLD_VERSIONS, new String[] {"vercode"}, " apkid=\""+apkid+"\" and ver=\""+ver+"\" ", null, null, null, null);
+			if(!c.moveToFirst()){
+				c.close();
+				return false;
+			}
+		}
+		
 		tmp.put("instver", ver);
-		tmp.put("instvercode", vercode);
+		tmp.put("instvercode", c.getInt(0));
 		c.close();
+		
 		return (db.insert(TABLE_NAME_LOCAL, null, tmp) > 0); 
 	}
 	
-	public boolean wasUpdate(String apkid, int versioncode){
-		Cursor c = db.query(TABLE_NAME, new String[] {"lastvercode"}, "apkid=\""+apkid+"\"", null, null, null, null);
-		c.moveToFirst();
+	public boolean wasUpdateOrDowngrade(String apkid, int versioncode){
+		Cursor c = db.query(TABLE_NAME_LOCAL, new String[] {"instvercode"}, "apkid=\""+apkid+"\"", null, null, null, null);
+		
+		if(!c.moveToFirst())
+			return true;
+		
 		int bd_code = c.getInt(0);
 		c.close();
-		return (versioncode == bd_code);
+		return (versioncode != bd_code);
 	}
 	
 	/*
@@ -307,7 +386,8 @@ public class DbHandler {
 		tmp.put("apkid", apkid);
 		tmp.put("instver", ver);
 		tmp.put("instvercode", vercode);
-		return (db.insert(TABLE_NAME_LOCAL, null, tmp) > 0); 
+		long i = db.insert(TABLE_NAME_LOCAL, null, tmp);
+		return i > 0; 
 	}
 	
 	public boolean UpdateInstalled(String apkid, String ver, int vercode){
@@ -386,19 +466,19 @@ public class DbHandler {
 				}else{
 					int instvercode = c.getInt(4);
 					int lastvercode = c.getInt(5);
-					if(instvercode == lastvercode){
-						node.status = 1;
-						node.ver = c.getString(2);
-					}else{
-						if(instvercode < lastvercode){
-							node.status = 2;
-							node.ver = c.getString(2) + "/ new: " + c.getString(3);
-						}else{
-							node.status = 1;
-							node.ver = c.getString(2);
-						}
-					}
 					
+					if(instvercode >= lastvercode){
+						
+						if(getOldApks(node.apkid).size()==0 && instvercode == lastvercode){
+							node.status = 1;
+						} else {
+							node.status = 3;
+						}
+						
+					}else{
+						node.status = 2;
+					}
+					node.ver = c.getString(2);
 				}
 				node.rat = c.getFloat(7);
 				node.down = c.getInt(8);
@@ -515,20 +595,20 @@ public class DbHandler {
 					node.status = 0;
 				}else{
 					//if(c.getString(2).equalsIgnoreCase(c.getString(3))){
+					node.ver = c.getString(2);
 					if(c.getInt(4) == c.getInt(5)){
 						node.status = 1;
-						node.ver = c.getString(2);
 					}else{
 						int instvercode = c.getInt(4);
 						int lastvercode = c.getInt(5);
 						if(instvercode < lastvercode){
 							node.status = 2;
-							node.ver = c.getString(2) + "/ new: " + c.getString(3);
+							node.ver += "/ new: " + c.getString(3);
 						}else{
 							node.status = 1;
-							node.ver = c.getString(2);
 						}
 					}
+					
 				}
 				node.rat = c.getFloat(7);
 				tmp.add(node);
@@ -546,7 +626,6 @@ public class DbHandler {
 	/*
 	 * Same function as above, used in list of updates
 	 */
-	
 	public Vector<ApkNode> getUpdates(String type){
 		Vector<ApkNode> tmp = new Vector<ApkNode>();
 		Cursor c = null;
@@ -614,20 +693,13 @@ public class DbHandler {
 		return tmp;
 	}
 	
-	
-	/*
-	 * Returned values about an application
-	 *  - vec(0): servers
-	 *  - vec(1): server version
-	 *  - vec(2): is it installed?
-	 *  - vec(3): installed version
-	 */
 	public Vector<String> getApk(String id){
 		Vector<String> tmp = new Vector<String>();
 		Cursor c = null;
 		int size = 0;
+		String lastvercode = "";
 		try{
-			c = db.query(TABLE_NAME, new String[] {"server", "lastver", "size"}, "apkid=\""+id.toString()+"\"", null, null, null, null);
+			c = db.query(TABLE_NAME, new String[] {"server", "lastver", "size", "lastvercode"}, "apkid=\""+id.toString()+"\"", null, null, null, null);
 			c.moveToFirst();
 			/*String tmp_serv = new String();
 			for(int i=0; i<c.getCount(); i++){
@@ -639,6 +711,9 @@ public class DbHandler {
 			tmp.add(c.getString(0));
 			tmp.add("\t" + c.getString(1)+"\n");
 			size = c.getInt(2);
+			
+			lastvercode = c.getInt(3)+"";
+			
 			c = db.query(TABLE_NAME_LOCAL, new String[] {"instver"}, "apkid=\""+id.toString()+"\"", null, null, null, null);
 			if(c.getCount() == 0){
 				tmp.add("\tno\n");
@@ -662,11 +737,12 @@ public class DbHandler {
 			
 			tmp.add(Float.toString(rat));
 			
-			if(size == 0)
+			if(size == 0){
 				tmp.add("Size: No information available");
-			else
+			}else{
 				tmp.add("Size: " + new Integer(size).toString() + "kb");
-			
+			}
+			tmp.add(lastvercode);
 			//c.close();
 		}catch (Exception e){
 			//System.out.println(e.toString());
@@ -676,12 +752,122 @@ public class DbHandler {
 		return tmp;
 	}
 	
+	/**
+	 * @author rafael
+	 * 
+	 * @param apk_id
+	 * @return
+	 */
+	public ArrayList<VersionApk> getOldApks(String apk_id){
+		ArrayList<VersionApk> tmp = new ArrayList<VersionApk>();
+		Cursor c = null;
+		try{
+			
+			c = db.query(TABLE_NAME_OLD_VERSIONS, new String[] {"ver", "size", "vercode"}, "apkid=\""+apk_id+"\"", null, null, null, null);
+			c.moveToFirst();
+			
+			do{
+				tmp.add( new VersionApk(c.getString(0), c.getInt(2), apk_id, c.getInt(1)) );
+			}while(c.moveToNext());
+			
+		}catch (Exception e){
+			//System.out.println(e.toString());
+		}finally{
+			c.close();
+		}
+		
+		return tmp;
+	}
 	
-	public Vector<DownloadNode> getPathHash(String id){
+	/**
+	 * @author rafael
+	 * 
+	 * @param apk_id
+	 * @return
+	 */
+	public ArrayList<VersionApk> getOldAndNewApks(String apk_id){
+		ArrayList<VersionApk> tmp = new ArrayList<VersionApk>();
+		tmp.addAll(getOldApks(apk_id));
+		
+		Cursor c = null;
+		try{
+			
+			c = db.query(TABLE_NAME, new String[] {"lastver", "size", "lastvercode"}, "apkid=\""+apk_id+"\"", null, null, null, null);
+			c.moveToFirst();
+			
+			do{
+				tmp.add(new VersionApk(c.getString(0),  c.getInt(2), apk_id, c.getInt(1)));
+			}while(c.moveToNext());
+			
+		}catch (Exception e){
+			//System.out.println(e.toString());
+		}finally{
+			c.close();
+		}
+		
+		return tmp;
+		
+	}
+	
+	
+	/**
+	 * @author rafael
+	 * 
+	 * @param apk_id
+	 * @param server
+	 * @return
+	 */
+	public Vector<String> copyFromRecentApkToOldApk(String apk_id, String server){
+		Vector<String> tmp = new Vector<String>();
+		Cursor c = null;
+		ApkNodeFull tmp_apk = new ApkNodeFull();
+		
+		tmp_apk.apkid = apk_id;
+		tmp_apk.name = "";
+		tmp_apk.ver = "0.0";
+		tmp_apk.vercode = 0;
+		tmp_apk.md5hash = "";
+		tmp_apk.path="";
+		tmp_apk.size = 0;
+		
+//		private static final String CREATE_TABLE_APTOIDE = "create table if not exists " + TABLE_NAME + " (apkid text, "
+//        + "name text not null, path text not null, lastver text not null, lastvercode number not null, "
+//        + "server text, md5hash text, size number default 0 not null, primary key(apkid, server));";
+//		private static final String CREATE_TABLE_EXTRA = "create table if not exists " + TABLE_NAME_EXTRA
+//		+ " (apkid text, rat number, dt date, desc text, dwn number, catg text default 'Other' not null,"
+//		+ " catg_ord integer default 2 not null, dt date, md5hash text, primary key(apkid));";
+		
+		try{
+			
+			c = db.query(TABLE_NAME, 
+					new String[] {"name", "path", "lastver", "lastvercode", "server", "md5hash", "size"}
+					, "server=\""+server+"\" and apkid=\""+apk_id+"\"", null, null, null, null);
+			c.moveToFirst();
+			
+			tmp_apk.name = c.getString(0);
+			tmp_apk.path = c.getString(1);
+			tmp_apk.ver = c.getString(2);
+			tmp_apk.vercode = c.getInt(3);
+			server = c.getString(4);
+			if(!c.isNull(5)) tmp_apk.md5hash = c.getString(5);
+			tmp_apk.size = c.getInt(6);
+			
+			this.insertOldApk(tmp_apk, server);
+			
+		}catch (Exception e){
+			//System.out.println(e.toString());
+		}finally{
+			c.close();
+		}
+		return tmp;
+	}
+	
+	
+	public Vector<DownloadNode> getPathHash(String id_apk, String ver){
 		Vector<DownloadNode> out = new Vector<DownloadNode>();
 		Cursor c = null;
 		try{
-			c = db.query(TABLE_NAME, new String[] {"server", "path", "md5hash", "size"}, "apkid='"+id.toString()+"'", null, null, null, null);
+			c = db.query(TABLE_NAME, new String[] {"server", "path", "md5hash", "size", "lastver"}, "apkid='"+id_apk+"' and lastver ='"+ver+"'", null, null, null, null);
 			c.moveToFirst();
 			for(int i =0; i<c.getCount(); i++){
 				String repo = c.getString(0);
@@ -692,6 +878,7 @@ public class DbHandler {
 				}
 				int size = c.getInt(3);
 				DownloadNode node = new DownloadNode(repo, remotePath, md5sum, size);
+				node.version = c.getString(4);
 				out.add(node);
 			}
 			//c.close();
@@ -702,6 +889,45 @@ public class DbHandler {
 		return out;
 	}
 	
+	/**
+	 * @author rafael
+	 * 
+	 * @param id_apk
+	 * @param ver
+	 * @return
+	 */
+	public Vector<DownloadNode> getPathHashOld(String id_apk, String ver){
+		
+		Vector<DownloadNode> out = new Vector<DownloadNode>();
+		Cursor c = null;
+		try{
+			c = db.query(TABLE_NAME_OLD_VERSIONS, new String[] {"server", "path", "md5hash", "size"}, "apkid='"+id_apk+"' and ver ='"+ver+"'", null, null, null, null);
+			c.moveToFirst();
+			for(int i =0; i<c.getCount(); i++){
+				String md5h = null;
+				String remotePath = c.getString(0)+"/"+c.getString(1);
+				if(!c.isNull(2)){
+					md5h = c.getString(2);
+				}
+				DownloadNode node = new DownloadNode(c.getString(0), remotePath, md5h, c.getInt(3));
+				node.version = ver;
+				out.add(node);
+			}
+			//c.close();
+		}catch(Exception e){
+		}finally{
+			c.close();
+		}
+		return out;
+		
+	}	
+	
+	/*
+	 * 
+	 * (0): server
+	 * (1): path
+	 * (2): md5hash (may not exist)
+	 */
 	public Vector<String> getPathHash2(String id){
 		Vector<String> out = new Vector<String>();
 		Cursor c = null;
@@ -731,10 +957,8 @@ public class DbHandler {
 			c.moveToFirst();
 			out = c.getString(0);
 			c.close();
-		}catch (Exception e){ }
-		finally{
-			c.close();
-		}
+		}catch (Exception e){}
+		finally{ c.close(); }
 		return out;
 	}
 	
@@ -995,5 +1219,8 @@ public class DbHandler {
 		Log.d("Aptoide","remved: " + del);
 		int a = db.delete(TABLE_NAME, "server='"+repo+"'", null);
 		Log.d("Aptoide","Removed: " + a);
+		int o = db.delete(TABLE_NAME_OLD_VERSIONS, "server='"+repo+"'", null);
+		Log.d("Aptoide","RemovedOld: " + o);
+		
 	}
 }
